Update CONTRIBUTING.md
Removed the 'Repository organization' section, because the information in it was incorrect. (I have a feeling it was copied from a contributor kit topic from another project maybe?) The topics are not organized by language. The topics are not named by MSDN GUID (since the topics have never been published on MSDN at all.)

thanks
AJ