Updating uset_get.md to include /users/ query
This file is supposed to show examples of "getting a user" and using the "/users/" endpoint. However, the only examples shown on this page use the /me endpoint...

I have modified the page to include a query example of the /users/ endpoint as well as the /me endpoint, and updated the $select example to also use the /users/ endpoint.