subscribe a resource on behalf of a user
Its not clear by the documentation how to subscribe to a given resource on behalf of a user in case of permissions given by administrator consent. Documentation provide examples that mention specific resource id or me/ resources. But in case of an administrator consent, you will be able to subscribe for notifications on a specific mailbox user, such as: john.doe@onmicrosoft.com.
In case that the resource isnt available you get an unclear error.