Update Application permissions
This article currently states that Application permissions are supported (Groups.ReadWrite.All). This isn't accurate as calling this API returns the same error that we expect when calling the other Conversation/ConversationThread APIs return when attempting to leverage Application permissions:

    'code': 'ErrorAccessDenied',
    'message': 'Access is denied. Check credentials and try again.',

I'm able to reproduce this on demand in my test tenant, and at least one customer can confirm this behavior as well. I'm confident that PG doesn't support Application permissions for this API, rather it is an error with the documentation.