Fix inverted boolean behavior for saveToSentItems
"By setting the `saveToSentItems` property to `true`" is wrong and the value must be FALSE as TRUE is the documented default and it makes no sense based on the property wording. If set to false it will not store a copy in the send items folder execpt an admin configured an exchange policy.