Add version info to outboundConnectionsRequired and outboundConnectionsBlocked
Intune was updated to only apply these settings to Windows versions 1809 and up so this updates the docs to reflect these changes. I also plan on making the changes in the service code so they get pulled in automatically, but wanted to make these changes now so they are in line with the service.