Fix incorrect statement about groups created via the azure portal
Only security groups and unified groups can be created via the azure portal. In order to have mailEnabled be true for both, the group types would have to be "mail-enabled security" and "unified". (I own the code in the azure portal for group creation and know for a fact that I set mailEnabled to false when creating security groups).