Correction to v1.0 getByIds documentation
The documentation for `getByIds` incorrectly states that full objects are returned. The returned User objects are not complete, though a workaround is possible using `$select` that supplies more complete User objects.
The beta version of this endpoint matches the documentation more closely and supplies more complete User objects without any workarounds.
So perhaps if the relevant parts of the beta version ever make it to v1.0, that would resolve the issue as well. In the meantime I think a documentation update for v1.0 would be helpful to other users of the Graph.

My previous attempt to get this corrected is #2487