Re-added changes from reverted commit
Had to revert e1556798de9bcd177425cfc90bb38c1d0e66ab47 to unblock merging master into live. The problem in that commit was a conversion of line endings in workbookcomment.md and workbookcommentreply.md, causing a merge conflict in the build system.

Re-adding the "real" changes from that commit (without the changes to the workbook*.md files)