Update application-proxy-configure-api.md
Series of updates to correct this flow to use the recommended pattern for app creation. This updates use Create Application instead of the custom application template and then also adds in additional steps for completing the configuration. Previous steps had a few issues that prevented the configuration from working end to end.