Explicit `userAgentApplication` passed to mgt
Re: https://github.com/microsoftgraph/microsoft-graph-toolkit/pull/576

Probably needs some wordsmithing. It's late, I'm tired. But this should hit the key points as a draft.

I wasn't sure on the best way to communicate the polymorphic nature of `MsalConfig`. It seems weird to me having the two interface declarations with the same name, but I didn't want to distract from the point either.

I also wasn't sure whether this repo supports `[!Note]` syntax. It didn't preview, and I can't find examples in the Graph docs, but one of those felt appropriate for the "be sure to understand collisions" paragraph.