Update teamstab.md
The Graph API /teams/{id}/channels/{id}/tabs started to return the teamsAppId with a value of null. Previously, it was possible to get the Teams App ID such as http://com.microsoft.teamspace.tab.wiki/ as part of the response. Since this property is deprecated, it might be helpful to provide developers with some guidance about the new approach.