Add sample code to retrieve IDs for permissions
As automation tools like Azure CLI, PowerShell, Terraform etc require you to use the ID instead of the name of the permission you want to use, it makes sense to have some sample code to show how to retrieve this. So I added some code that shows how to do this, as it isn't the simplest thing to figure out in my opinion. It took me a while to figure out, And I can't find the IDs documented anywhere.