Update user-list-manager.md - Added requirement
An essential requirement when trying to select properties of an 
expanded manager property is to add the $count=true parameter
somewhere in the query. Example 2 exhibits this; however, it is not
clearly indicated in the documentation. I added a note that 
highlights this. In addition, I emphasized the need for the required 
header: ConsistencyLevel eventual. Although this is documented, 
our customers frequently miss this requirement, and additional 
emphasis should be helpful. The following note was added:
"In order to select the expanded manager's properties, the 
`$count=true` parameter must be added to the query as well 
as the header, `ConsistencyLevel=eventual`. You can see this 
implemented in Example 2 below."