Clarity/Correction needed for note about a single App Access Policy affected all MS Graph/EWS Apps
The following note needs to be adjusted (or removed entirely, which I propose as the note is inaccurate and misleading):

> [!NOTE]
> Once an application access policy is configured for an organization, the policy applies to any app that uses the Microsoft Graph API or Exchange Web Services to access Exchange Online mailboxes.


This isn't accurate. Creating a single Application Access Policy does not affect all MS Graph/EWS applications - it only affects applications using the App Id that the Application Access Policy was created for (as creating a new App Access Policy requires a specific App Id to be provided at creation). The current phrasing leads customers to believe that creating a single Application Access Policy (that's tied to a specific App Id) will affect ALL App Registrations.