add expand of assignment target and resource roles
In the EM example to retrieve access package assignments, the description said the call could be used "...to see that resources have been assigned to the Requestor1 user account."  but the $expand to include that in the result was inadvertantly omitted.  This change adds the $expand for target and resource roles and illustrates those being included in the GET response.