Update known-issues.md
Updating Known Issues section for allowExternalSenders property to document that this property can only be used for Unified Groups in Microsoft Graph. If an attempt is made to access the allowExternalSenders property on a Distribution List/Security Group (including GET operations), Microsoft Graph will return a MailboxNotEnabledForRESTAPI which is by design.