Per request, adding my edits to the Beta file as well
The ask was to keep files in Sync, by adding my edits to this beta file.  The original file received additions via the PR review process to the description and title which I have not added to this Beta file, as they are not my edits. Because other parts of these files are not in sync I am not comfortable editing the title and description of this beta file. If those additional edits are needed, as part of the PR processing that added them to my original edit please add those to this beta file as well.