Delete Role Assignable Groups
adding a note for role assignable group deletion. Role assignable groups can apparently only be deleted by Users who created the group or are a global admin or privileged role admin. This is documented internally in a troubleshooting wiki, can be validated using DS explorer, and is mentioned under the Group page for MS Graph but isn't mentioned on the MS Graph Delete Group documentation.

The reasoning for this is if I as a customer attempted to make a request to delete a group and a 403 was returned, a reasonable first instinct would be to check the documentation for the deletion action first, rather than the overview page defining everything about a group. Since this is a side effect caused by the property being set, it wouldn't be immediately clear that this property in the cause.