Add note regarding permissions
When customer make a call to modify group member, they also need the read and write permission to the member object. If it's adding a user then the app needs User.ReadWrite.All, if it's adding a device, the app needs Device.ReadWrite.All. Same as service principals. This part is not stated clearly in the doc so partner team has difficulties in granting those high privileged permission for their first party apps as these requirements for users, devices, service principals are not stated in the doc. 