Update getretentionlabel.md
Fixed this 

Line 17, Column 86:  [Suggestion] You've pinned this link to a specific version of content with the view parameter. It's recommended not to pin a version unless that version is A) not the default view and B) the context is about that version specifically. To proceed with pinning a version add the &preserve-view=true to the URL. Otherwise, remove the view parameter. URL: /microsoft-365/compliance/auto-apply-retention-labels-scenario?view=o365-worldwide