Removing ".Shared" requirement
As per confirmation from engineering, this API does not support the scenario where "a user requests time suggestions on behalf of another, even if they have delegate permissions for said user's calendar". As such, correcting the issue of this doc connotating that the use of the user-delegation ".Shared" permission is required.