Adjusting for customer flow if they land on API page first
The docs are organized such that we expect folks to start with the "learning" sections of our docs if they want to learn how to use our APIs.  However, google search will bring people first to the API page.  So this is tweaked to help folks find the right resources.