Update application-post-tokenlifetimepolicies.md
Purpose: The purpose of this change is to make the documentation more clear while Assigning a tokenLifetimePolicy to an application using Graph-API.

Impact: This change will bring more clarity into the documentation and make it up-to-date. This will make it easier for developers to use the API, which will lead to increased adoption of the API.

Implementation: The change will be implemented by updating the existing documentation and adding one new line as needed.