beta and v1 template files
Fixing for this comment:
GLOBAL:
I'm using this example - the create operation for citationTemplate - but it looks like all the create operation for the 6 template resources have the similar issue.
Issue:
In lines 43-49, you only list 3 properties. Is setting the createdBy property necessary, which you show in the example request in line 73?
Action for you:
(a) Please confirm if specifying the createdBy property is necessary in this operation.
If it's not necessary, please remove the createdBy property in the example request body of the create operation topic.
(b) Please clarify in the resource topic for citationTemplate, in the Properties table, the descriptions for createdBy and createdDateTime, add to the end of the existing description "Read-only." (That would be after the sentence that in another comment I requested you to fix, "Inherited from microsoft.graph.security.filePlanDescriptorTemplate.").
(c) And just to make it consistent for the properties in that Properties table, add the same "Read-only." sentence to the end of the description of the id property as well. Just so we are clear, in general, the id property of all the resources of type entity are read-only in Microsoft Graph, assigned on creation. Since we mark one of the properties in the same table as "read-only" in this case, I don't want to create ambiguity that all the other properties in the same table are not read-only if they are not marked as such. :-)
Please do (a) to (c) for all the 6 template resources, for each of v1 and beta.