Make it more explicit that the "nextSigningCertificate" is also accepted at anytime
Previously, we could understand that the "nextSigningCertificate" was only usable when the primary "signingCertificate" was expired. But, even though that's the main use case, actually both certificates are always accepted as token signers!

I confirmed this by doing some tests and I reported this to MSRC (VULN-113138 CRM:0632004158) who considered it normal. So I hope this doc change makes it clearer.