Update permissions-grant-via-msgraph.md
You will be able to see the new permission at the Enterprise Applications - PermissionsTab
Under the App registration it will be Other Permissions add at the end of the list for API permissions tab. In order to move that permission and activate you need to Grant Admin Consent.



I was following this guide for python and I found a little bit challenge. So adding the things I believe it would make it more clear
1) The object ID for the SP is to be taken from Enterprise Applications.

2) The user that will pass the credentials to alter the SP must have high permissions to do it so. I was able to do that with one SP added as global administrator and that one given permissions to a second SP.
3) Which libraries for python to be able to use the sample code.

This was my personal step by step:

1 Replace the information client_ID and Client_secret with the SP with high permissions. This first step will represent the credential that must have enough permissions for the next step.


tenant_id = "Replace with Tenant ID"
client_id = "Replace with Service Principal ID that has high permissions"
client_secret = "Replace with secret from the SP that has high permissions" 

6.2 The next step you will replace with the Service Principal that you will change the permissions. For example adding those 2 permissions. You will use the Object ID as Client_ID that you got from the Enterprise Applications:
•	User.Read.All
•	User.ReadWrite.All 
In the code context:
•	Resource ID is the Microsoft Graph ID that
•	Client_ID is the Object ID from Enterprise Application
•	App Role id is the ID of the role that you want to change