Updating parameters required to create a local B2C identity
This is no longer the case as tested in the lab: "For a B2C local account identity, only passwordProfile is required, and passwordPolicies must be set to DisablePasswordExpiration."

To create a B2C local account identity, passwordProfile, accountEnabled, displayName, mailNickname, userPrincipalName are required. Only including passwordProfile will results in a 400 (bad request) error. 

Because most of the parameters listed on the table are required, suggest removing the sentence (as opposed to listing the required parameters).