Update insights-customize-item-insights-privacy.md
The reason for this change? We have evidence from an ongoing investigation/critsit with Charter that 24 hours is simply wrong. We are current at 72 hours and no hope in sight. 
 
We believe that the time to update insights setting is based on the size of the body of content within the entire tenant… but I can’t think of a good way to convert that into a number of hours. Better to simply delete that line from the documentation in my opinion.