Update identity-governance-pim-rules-overview.md
In the "Activation rules" section, the setting marked as #2 seems to be mapped to the incorrect rule. Instead of "Enablement_Admin_Eligibility", it should be "Enablement_EndUser_Assignment". The 'Enforced for caller' column should likely also be updated to 'End User'.

This is further supported by a separate Learn page which shows how to correctly call this API: https://learn.microsoft.com/en-us/graph/how-to-pim-update-rules?tabs=http#example-2-update-the-justification-mfa-and-ticketing-rules-required-on-activation

For more context/repros, please see my confusion (and seeming resolution) on my Q&A post here:
https://learn.microsoft.com/en-us/answers/questions/1656721/pim-activation-settings-for-an-entra-role-exposed