Update preauthorizedapplication.md
The following public document states that "In some rare cases, an identifier listed in the delegatedPermissionIds property may actually identify an app role (from the service principal's appRoles property), indicating that the client application identified by the appId property has been preauthorized for that app role." https://learn.microsoft.com/en-us/graph/api/resources/preauthorizedapplication?view=graph-rest-1.0

It is recommended that the following statement be added to ensure that customers don't think we support preauthorization of app roles to clients. The app role technically is modelled as an app-only scope.  Pre-authorization of app-only scopes to clients is not currently supported. Validation checks are performed at runtime to ensure that the current PermissionIds are of type DelegatedPermissions. In scenarios where customers want to avoid the Consent experience associated with app roles, they should grant AppRoles to a client app programmatically. Reference: Grant an appRoleAssignment to a service principal - Microsoft Graph v1.0 | Microsoft Learn