Update changelog for new protocols/flows in sign in logs
We are extending the existing `authenticationProtocol` enum in `auditLogs/signIns` with some new protocol/flow values. 

This update does not change the existing enum structure, and there's no dedicated document for the `authenticationProtocol` enum, so there are no changes to the actual documentation. 

The only thing we need to update, IIUC, is the changelog.