sites/siteID/onenote/sectiongroups will list folders that are not notebook folders
the article is misleading. is suggests that I could run sites/siteID/onenote/sectionsgroups and it would automatically filter for all notebook sectiongroups in the site. instead it present all subfolders of the site pages root folder, including all other folders. so as the endpoint is not filtering for notebooks the filter needs to be added manuallys. I do not see any use case where customers whould not have the goal to filter for notebook subfolders/sectiongroups we are getting regular cases with this confusion.