Update joinedTeams and teams responses to be aligned...
with the returned responses.
The known issue mentioned for joinedTeams response is actually
supposed to be just a note like the one in list teams response.

Responses for these APIs are redacted most probably for saving
bandwidth. The complete responses can be fetched from /teams/{id}
API.

Updated the response objects to look similar to what a user would
see.